requirements.txt
streamlit
pymongo
pandas
python-dotenv
plotly